+++ deps/uv/src/unix/netbsd.c
@@ -82,7 +82,7 @@ int uv_exepath(char* buffer, size_t* size) {
   return 0;
 }
 
-double uv_get_free_memory(void) {
+uint64_t uv_get_free_memory(void) {
   struct uvmexp info;
   size_t size = sizeof(info);
   int which[] = {CTL_VM, VM_UVMEXP};
@@ -91,10 +91,10 @@ double uv_get_free_memory(void) {
     return -1;
   }
 
-  return (double) info.free * psysconf(_SC_PAGESIZE);
+  return (uint64_t) info.free * psysconf(_SC_PAGESIZE);
 }
 
-double uv_get_total_memory(void) {
+uint64_t uv_get_total_memory(void) {
 #if defined(HW_PHYSMEM64)
   uint64_t info;
   int which[] = {CTL_HW, HW_PHYSMEM64};
@@ -108,5 +108,5 @@ double uv_get_total_memory(void) {
     return -1;
   }
 
-  return (double) info;
+  return (uint64_t) info;
 }
